% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNumberToImage.R
\name{addNumberToImage}
\alias{addNumberToImage}
\alias{addnumbertoimage}
\alias{addNumbertoimage}
\alias{addNumberToimage}
\alias{addnumberToImage}
\alias{addnumbertoImage}
\title{addNumberToImage}
\usage{
addNumberToImage(
  image = NULL,
  number = NULL,
  pos_x = NULL,
  pos_y = NULL,
  number_size_factor = NULL,
  number_color = "red"
)
}
\arguments{
\item{image}{An three-dimensional array of numbers between 0 and 1}

\item{number}{A number (integer to be drawn/copied)}

\item{pos_x}{A number (x-value for starting pixel (upper left corner))}

\item{pos_y}{A number (y-value for starting pixel (upper left corner))}

\item{number_size_factor}{A number (factor for resizing the number)}

\item{number_color}{A string (color of thr number to be drawn)}
}
\description{
Adds a number to the image layer
}
\details{
Adds an integer to a x-y-3(rgb)-representation of an image.
}
\author{
Kai Budde
}
