% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addScaleBar.R
\name{addScaleBar}
\alias{addScaleBar}
\alias{addscaleBar}
\alias{addScalebar}
\alias{addscalebar}
\title{addScaleBar}
\usage{
addScaleBar(
  image = NULL,
  length_per_pixel = NULL,
  distance_from_border = 20,
  number_size_factor = 1
)
}
\arguments{
\item{image}{An array (2 or 3 dimensional array, could also be of
formal class Image by package EBImage)}

\item{number_size_factor}{A number (factor for resizing the number)}
}
\description{
Add a horizontal scale bare to image
}
\details{
Add a scale bar to a tiff image depending on the meta data of
the czi file
}
\author{
Kai Budde
}
