% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellPixels.R
\name{cellPixels}
\alias{cellPixels}
\alias{cellpixels}
\alias{CellPixels}
\title{cellPixels}
\usage{
cellPixels(input_dir = NULL, nucleus_color = "blue", number_size_factor = 0.1)
}
\arguments{
\item{input_dir}{A character (directory that contains all images)}
}
\description{
Main function for counting pixels in different regions
}
\details{
Input should be tif-format.
We are counting the intensities of pixels of different color within and
outside of the nuclei. We also add information about the entire image.
}
\author{
Kai Budde
}
