% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellSpreading.R
\name{cellSpreading}
\alias{cellSpreading}
\alias{cellspreading}
\alias{CellSpreading}
\alias{Cellspreading}
\title{cellSpreading}
\usage{
cellSpreading(
  input_dir = NULL,
  nucleus_color = "blue",
  cell_body_color = "red",
  number_of_pixels_at_border_to_disregard = 3,
  use_histogram_equalized = FALSE,
  normalize_nuclei_layer = FALSE,
  magnification_objective = NULL,
  apotome_section = FALSE,
  blur_sigma = NULL,
  thresh_w_h_nuc = NULL,
  thresh_offset = NULL,
  number_size_factor = 1,
  add_scale_bar = FALSE
)
}
\arguments{
\item{input_dir}{A character (directory that contains all images)}

\item{nucleus_color}{A character (color (layer) of nuclei)}

\item{cell_body_color}{A character (color (layer) of cell body)}

\item{number_of_pixels_at_border_to_disregard}{A number (number of pixels
at the border of the image (rows and columns) that define the region
where found cells are disregarded)}

\item{use_histogram_equalized}{A boolean (use histogram equalized images for everything)}

\item{normalize_nuclei_layer}{A boolean (state whether nucleus layer should be normalized)}

\item{magnification_objective}{A number (magnification of objective if not given in metadata or if metadata is wrong)}

\item{apotome_section}{A boolean (TRUE is sectioned image shall be used)}

\item{blur_sigma}{A number (blurring factor)}

\item{thresh_w_h_nuc}{A number (width and heigth for thresholding)}

\item{thresh_offset}{A number (offset for thresholding)}

\item{number_size_factor}{A number (factor to resize numbers for
numbering nuclei)}

\item{add_scale_bar}{A logic (add scale bar to all images that are saved
if true)}
}
\description{
Main function for counting pixels in different regions
}
\details{
Input should be czi or tif-format with dim(z)>=1.
We are trying to identify different cells by using nuclei and Actin layers.
}
\author{
Kai Budde
}
