% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLayer.R
\name{getLayer}
\alias{getLayer}
\alias{getlayer}
\title{getLayer}
\usage{
getLayer(image = NULL, layer = NULL)
}
\arguments{
\item{image}{An three-dimensional array of numbers between 0 and 1}

\item{layer}{A character (color of the layer)}
}
\description{
Get a specific layer of the image
}
\details{
By using an x-y-3(rgb)-representation of an image, you can
extract one layer.
}
\author{
Kai Budde
}
