% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeIntensity.R
\name{normalizeIntensity}
\alias{normalizeIntensity}
\alias{normalizeintensity}
\title{normalizeIntensity}
\usage{
normalizeIntensity(image = NULL)
}
\arguments{
\item{image}{An three-dimensional array of numbers between 0 and 1}
}
\description{
Normalize the intensity of the image
}
\details{
Normalize the intensity of an image by dividing all values by
the maximum intensity.
}
\author{
Kai Budde
}
