% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellPixels.R
\name{cellPixels}
\alias{cellPixels}
\alias{cellpixels}
\alias{CellPixels}
\title{cellPixels}
\usage{
cellPixels(
  input_dir = NULL,
  nucleus_color = "blue",
  protein_in_nuc_color = NULL,
  number_size_factor = 0.2,
  bit_depth = NULL
)
}
\arguments{
\item{input_dir}{A character (directory that contains all images)}

\item{nucleus_color}{A character (color (layer) of nuclei)}

\item{protein_in_nuc_color}{A character (color (layer) of protein
expected in nucleus)}

\item{number_size_factor}{A number (factor to resize numbers for
numbering nuclei)}

\item{bit_depth}{A number (bit depth of the original czi image)}
}
\description{
Main function for counting pixels in different regions
}
\details{
Input should be tif-format.
We are counting the intensities of pixels of different color within and
outside of the nuclei. We also add information about the entire image.
}
\author{
Kai Budde
}
