% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellPixels.R
\name{cellPixels}
\alias{cellPixels}
\alias{cellpixels}
\alias{CellPixels}
\title{cellPixels}
\usage{
cellPixels(
  input_dir = NULL,
  nucleus_color = "blue",
  protein_in_nuc_color = "none",
  protein_in_cytosol_color = "none",
  number_size_factor = 0.2,
  bit_depth = NULL,
  number_of_pixels_at_border_to_disregard = 3,
  add_scale_bar = FALSE,
  thresh_w_h_nuc = NULL,
  thresh_offset = NULL,
  blur_sigma = NULL,
  use_histogram_equalized = FALSE
)
}
\arguments{
\item{input_dir}{A character (directory that contains all images)}

\item{nucleus_color}{A character (color (layer) of nuclei)}

\item{protein_in_nuc_color}{A character (color (layer) of protein
expected in nucleus)}

\item{protein_in_cytosol_color}{A character (color (layer) of protein
expected in cytosol)}

\item{number_size_factor}{A number (factor to resize numbers for
numbering nuclei)}

\item{bit_depth}{A number (bit depth of the original czi image)}

\item{number_of_pixels_at_border_to_disregard}{A number (number of pixels
at the border of the image (rows and columns) that define the region
where found cells are disregarded)}

\item{add_scale_bar}{A logic (add scale bar to all images that are saved
if true)}
}
\description{
Main function for counting pixels in different regions
}
\details{
Input should be tif-format.
We are counting the intensities of pixels of different color within and
outside of the nuclei. We also add information about the entire image.
}
\author{
Kai Budde
}
