% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellPixels.R
\name{cellPixels}
\alias{cellPixels}
\alias{cellpixels}
\alias{CellPixels}
\title{cellPixels}
\usage{
cellPixels(
  input_dir = NULL,
  apotome = FALSE,
  apotome_section = FALSE,
  nucleus_color = "blue",
  min_nucleus_size = NULL,
  min_cytosol_size = NULL,
  protein_in_nucleus_color = "none",
  protein_in_cytosol_color = "none",
  protein_in_membrane_color = "none",
  number_size_factor = 0.2,
  bit_depth = NULL,
  number_of_pixels_at_border_to_disregard = 3,
  add_scale_bar = FALSE,
  thresh_w_h_nucleus = NULL,
  thresh_w_h_cytosol = NULL,
  thresh_w_h_protein_in_nucleus = NULL,
  thresh_offset_nucleus = NULL,
  thresh_offset_protein_in_nucleus = NULL,
  thresh_offset_protein_in_cytosol = NULL,
  blur_sigma = NULL,
  use_histogram_equalized = FALSE,
  metadata_file = NULL,
  normalize_nuclei_layer = FALSE,
  magnification_objective = NULL
)
}
\arguments{
\item{input_dir}{A character (directory that contains all images)}

\item{apotome}{A boolean (TRUE if Apotome was used)}

\item{apotome_section}{A boolean (TRUE is sectioned image shall be used)}

\item{nucleus_color}{A character (color (layer) of nuclei)}

\item{min_nucleus_size}{A number (minimum size in pixels of nuclei to be kept)}

\item{min_cytosol_size}{A number (minimum size in pixels of cytosol to be kept)}

\item{protein_in_nucleus_color}{A character (color (layer) of protein
expected in nucleus)}

\item{protein_in_cytosol_color}{A character (color (layer) of protein
expected in cytosol)}

\item{protein_in_membrane_color}{A character (color (layer) of protein
expected in membrane)}

\item{number_size_factor}{A number (factor to resize numbers for
numbering nuclei)}

\item{bit_depth}{A number (bit depth of the original czi image)}

\item{number_of_pixels_at_border_to_disregard}{A number (number of pixels
at the border of the image (rows and columns) that define the region
where found cells are disregarded)}

\item{add_scale_bar}{A logic (add scale bar to all images that are saved
if true)}

\item{thresh_w_h_nucleus}{A number (width and height of moving rectangle for
nucleus detection)}

\item{thresh_w_h_cytosol}{A number (width and height of moving rectangle for
cytosol detection)}

\item{thresh_w_h_protein_in_nucleus}{A number (width and height of moving
rectangle for protein in nucleus detection (different from detecting
nuclei themselves))}

\item{thresh_offset_nucleus}{A number (threshold for finding nuclei)}

\item{thresh_offset_protein_in_nucleus}{A number (threshold for
determining "positive" cells containing a certain protein in nucleus area)}

\item{thresh_offset_protein_in_cytosol}{A number (threshold for
determining "positive" cells containing a certain protein in cytosol area)}

\item{metadata_file}{A character (file with meta data if tifs are used)}

\item{normalize_nuclei_layer}{A boolean (state whether nucleus layer should be normalized)}

\item{magnification_objective}{A number (magnification of objective if not given in metadata or if metadata is wrong)}
}
\description{
Main function for counting pixels in different regions
}
\details{
Input should be czi or tif-format with dim(z)>=1.
We are counting the intensities of pixels of different color within and
outside of the nuclei. We also add information about the entire image.
}
\author{
Kai Budde
}
